Allow building with singletons-2.7 (GHC 8.10)
`singletons-2.7` now generates type families with
`StandaloneKindSignatures`, which are only available on GHC 8.10 or
later. I've enabled that extension additionally in the `.cabal` file
along with `CUSKs`, to prevent the new kind inference algorithm that
`StandaloneKindSignatures` uses from messing up other declarations
in this repo that do not have standalone kind signatures.